Generative_AI_for_Data/Fall_2024/GROUP_18-Formula_1_Regulations_RAG_System/Paddock-Pal/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
apache-airflow==2.10.3
apache-airflow-providers-common-compat==1.2.2
apache-airflow-providers-common-io==1.4.2
apache-airflow-providers-common-sql==1.20.0
apache-airflow-providers-fab==1.5.1
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-http==4.13.3
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-smtp==1.8.1
apache-airflow-providers-sqlite==3.9.1
apispec[yaml]==6.8.0
argcomplete==3.5.2
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
babel==2.16.0
blinker==1.9.0
boto3==1.35.80
botocore==1.35.80
cachelib==0.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
configupdater==3.2
connexion[flask]==2.14.2
cron-descriptor==1.4.5
croniter==5.0.1
cryptography==44.0.0
dataclasses-json==0.6.7
datetime==5.5
deprecated==1.2.15
dill==0.3.9
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.6
filelock==3.16.1
flask==2.2.5
flask-appbuilder==4.5.2
flask-babel==2.0.0
flask-caching==2.3.0
flask-jwt-extended==4.7.1
flask-limiter==3.9.2
flask-login==0.6.3
flask-session==0.5.0
flask-sqlalchemy==2.5.1
flask-wtf==1.2.2
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
gitpython==3.1.43
google-re2==1.1.20240702
googleapis-common-protos==1.66.0
grpcio==1.68.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
inflection==0.5.1
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.11
langchain-community==0.3.11
langchain-core==0.3.24
langchain-huggingface==0.1.2
langchain-pinecone==0.2.0
langsmith==0.2.3
lazy-object-proxy==1.10.0
limits==3.14.1
lockfile==0.12.2
mako==1.3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.23.1
mdit-py-plugins==0.4.2
methodtools==0.4.7
multidict==6.1.0
numpy==1.26.4
openai==0.28.0
pandas==2.2.3
pathspec==0.12.1
pendulum==3.0.0
pinecone-client==5.0.1
platformdirs==4.3.6
protobuf==5.29.1
psutil==6.1.0
pydantic==2.10.3
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.6.0
six==1.17.0
snowflake-connector-python==3.12.4
sqlalchemy==1.4.54
sqlparse==0.5.3
streamlit==1.41.0
tenacity==9.0.0
tqdm==4.67.1
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
watchdog==6.0.0
werkzeug==2.2.3
wrapt==1.17.0
yarl==1.18.3
